refactor: move mode types into mode.go
